order_api/requirements.txt
flask
mysql-connector-python
pika